# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Node Comics API",
    "description": "私有漫画图片、持久化翻译任务、普通与 PLUS 会员权益及周期页数额度。",
    "version": "0.3.0"
  },
  "paths": {
...
    "/v1/translation-submissions": {
      "post": {
        "tags": [
          "submissions"
        ],
        "summary": "Submit",
        "operationId": "submit_v1_translation_submissions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "submissions"
        ],
        "summary": "Submissions",
        "operationId": "submissions_v1_translation_submissions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/auth/dev": {
      "post": {
        "summary": "Dev Login",
        "operationId": "dev_login_v1_auth_dev_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/images/{asset_id}/access": {
      "get": {
        "summary": "Image Access",
        "operationId": "image_access_v1_images__asset_id__access_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/jobs/{job_id}": {
      "get": {
        "summary": "Job Get",
        "operationId": "job_get_v1_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessResponse": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "authorization_required": {
            "type": "boolean",
            "title": "Authorization Required"
          }
        },
        "type": "object",
        "required": [
          "url",
          "expires_at",
          "authorization_required"
        ],
        "title": "AccessResponse"
      },
...
      "DevLogin": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "pattern": "^[\\w\\-\\u4e00-\\u9fff ]+$",
            "title": "Username"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username"
        ],
        "title": "DevLogin"
      },
...
      "ErrorInfo": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorInfo"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ItemRequest": {
        "properties": {
          "client_item_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Client Item Id"
          },
          "image_sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Image Sha256"
          },
          "byte_size": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Byte Size"
          },
          "content_type": {
            "type": "string",
            "maxLength": 60,
            "title": "Content Type"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name",
            "default": ""
          },
          "file_hash": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-f0-9]{64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Hash"
          },
          "page_index": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Index"
          },
          "asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "client_item_id",
          "image_sha256",
          "byte_size",
          "content_type"
        ],
        "title": "ItemRequest"
      },
...
      "JobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "input_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Asset Id"
          },
          "image_sha256": {
            "type": "string",
            "title": "Image Sha256"
          },
          "file_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Hash"
          },
          "page_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Index"
          },
          "change_sequence": {
            "type": "integer",
            "title": "Change Sequence",
            "default": 0
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "preload"
          },
          "requested_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Asset Id"
          },
          "output_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Asset Id"
          },
          "result_available": {
            "type": "boolean",
            "title": "Result Available"
          },
          "result_expired": {
            "type": "boolean",
            "title": "Result Expired"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "target_language": {
            "type": "string",
            "title": "Target Language"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "quota_pages": {
            "type": "integer",
            "title": "Quota Pages"
          },
          "quota_kind": {
            "type": "string",
            "title": "Quota Kind"
          },
          "quota_period_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quota Period Id"
          },
          "settlement": {
            "type": "string",
            "title": "Settlement"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "cache_hit": {
            "type": "boolean",
            "title": "Cache Hit"
          },
          "reused": {
            "type": "boolean",
            "title": "Reused",
            "default": false
          },
          "submission_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Id"
          },
          "ordinal": {
            "type": "integer",
            "title": "Ordinal"
          },
          "cancel_requested": {
            "type": "boolean",
            "title": "Cancel Requested"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "quality_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Quality Flags"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "input_asset_id",
          "image_sha256",
          "output_asset_id",
          "result_available",
          "result_expired",
          "mode",
          "target_language",
          "status",
          "phase",
          "quota_pages",
          "quota_kind",
          "quota_period_id",
          "settlement",
          "version",
          "cache_hit",
          "ordinal",
          "cancel_requested",
          "error",
          "quality_flags",
          "created_at",
          "completed_at"
        ],
        "title": "JobResponse"
      },
...
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "user"
        ],
        "title": "LoginResponse"
      },
...
      "SubmissionRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "classic",
              "redraw"
            ],
            "title": "Mode"
          },
          "target_language": {
            "type": "string",
            "title": "Target Language"
          },
          "max_quota_pages": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Max Quota Pages"
          },
          "expected_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Kind"
          },
          "regenerate": {
            "type": "boolean",
            "title": "Regenerate",
            "default": false
          },
          "rerun_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rerun Job Id"
          },
          "acknowledge_unknown_cost": {
            "type": "boolean",
            "title": "Acknowledge Unknown Cost",
            "default": false
          },
          "reading_session_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading Session Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemRequest"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "target_language",
          "max_quota_pages",
          "items"
        ],
        "title": "SubmissionRequest"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "role"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}